+++ src/plaid/bridges/huggingface_helpers.py
@@ -7,10 +7,11 @@
 #
 #
 
-import pickle
 from pathlib import Path
+from typing import Union
 
 import datasets
+from datasets import load_from_disk
 
 from plaid import Sample
 from plaid.bridges.huggingface_bridge import to_plaid_sample
@@ -19,12 +20,11 @@ from plaid.bridges.huggingface_bridge import to_plaid_sample
 class _HFToPlaidSampleConverter:
     """Class to convert a Hugging Face dataset sample to a plaid :class:`Sample <plaid.containers.sample.Sample>`."""
 
-    def __init__(self, ds: datasets.Dataset):
-        self.ds = ds
+    def __init__(self, hf_ds: Union[datasets.Dataset, datasets.DatasetDict]):
+        self.hf_ds = hf_ds
 
     def __call__(self, sample_id: int) -> Sample:  # pragma: no cover
-        data = pickle.loads(self.ds[sample_id]["sample"])
-        return to_plaid_sample(data)
+        return to_plaid_sample(self.hf_ds[sample_id])
 
 
 class _HFShardToPlaidSampleConverter:
@@ -36,9 +36,9 @@ class _HFShardToPlaidSampleConverter:
         Args:
             shard_path (Path): path of the shard.
         """
-        from datasets import load_from_disk
-
-        self.hf_ds: datasets.Dataset = load_from_disk(shard_path.as_posix())
+        self.hf_ds: Union[datasets.Dataset, datasets.DatasetDict] = load_from_disk(
+            shard_path.as_posix()
+        )
 
     def __call__(
         self, sample_id: int