+++ src/plaid/containers/sample.py
@@ -54,10 +54,6 @@ from plaid.utils.base import safe_len
 from plaid.utils.deprecation import deprecated
 
 logger = logging.getLogger(__name__)
-logging.basicConfig(
-    format="[%(asctime)s:%(levelname)s:%(filename)s:%(funcName)s(%(lineno)d)]:%(message)s",
-    level=logging.INFO,
-)
 
 
 class Sample(BaseModel):